requirements.txt
langchain
langchain-community
langchain_experimental
pydantic
python-magic
fastapi
uvicorn
transformers
sentence-transformers
faiss-cpu
torch
ollama
colorlog
mypy
pytest
black
isort
flake8
streamlit
argparse
pyyaml
pygit
qdrant-client
gitpython
tensorflow
psutil
GPUtil
tf-keras